pyproject.toml
[project]
name = "backer"
version = "0.7.2"
description = "📦 Continuously back up files 📦"
authors = [{ name = "Tom Ritchford", email = "tom@swirly.com" }]
requires-python = ">=3.10"
readme = "README.md"
license = "MIT"
classifiers = ["Programming Language :: Python :: 3", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Programming Language :: Python :: 3.13", "Programming Language :: Python :: 3.14"]
dependencies = [
    "python-dotenv>=0.21.1,<0.22",
    "pyyaml>=6.0,<7",
    "run-subprocess>=0.9.1,<0.10",
    "safe-writer>=0.9.3,<0.10",
    "schedule>=1.1.0,<2",
    "watchdog>=2.2.1,<3",
]

[dependency-groups]
dev = [
    "coverage>=7.1.0,<8",
    "flake8==5.0.4",
    "impall>=1.1.1,<2",
    "gitz>=1.4.1,<2",
    "pytest>=7.2.1,<8",
    "pytest-benchmark>=4.0.0,<5",
    "pyupgrade>=3.21.2",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]

[tool.uv]

[tool.uv.build-backend]
module-root = ""


[tool.ruff]
[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
[tool.coverage.run]
branch = true
source = "backer/"

[tool.coverage.report]
fail_under = "50"
skip_covered = true
exclude_lines = ["pragma: no cover", "if False:", "if __name__ == .__main__.:", "raise NotImplementedError"]

